extract QuestionInput from Question
