+++ src/components/FormTextAreaInput/FormTextAreaInput.module.scss
@@ -13,8 +13,7 @@
   box-sizing: border-box;
 
   &:hover,
-  &:focus,
-  &:focus-within {
+  &:focus {
     border: 1px solid rgb(0, 143, 180);
   }
 }
