+++ src/pages/Browse/TaskDetail/Question/QuestionInput/QuestionInput.module.scss
@@ -1,0 +1,60 @@
+.avatar_wrapper{
+  width: 40px;
+  height: 40px;
+  display: block;
+  margin-right: 10px;
+}
+
+.input{
+  box-sizing: border-box;
+  display: table-cell;
+  margin: 0;
+  position: relative;
+  vertical-align: top;
+  width: 100%;
+  padding: 8px;
+  box-shadow: inset 0 1px 2px 0 hsla(0,0%,75.3%,.5);
+  border: 1px solid #bbc2dc;
+  border-radius: 4px;
+  cursor: text;
+  overflow: hidden;
+
+  textarea{
+    border: none;
+    height: 80px;
+    width: 100%;
+    margin: 0;
+    padding: 0;
+
+    &:focus,
+    &:hover{
+      border: none;
+    }
+  }
+
+  .input_footer{
+    display: flex;
+    flex-direction: row-reverse;
+    align-items: center;
+
+    .word_count{
+      display: block;
+      color: #bbc2dc;
+      font-size: 13px;
+    }
+
+    .submit_button{
+      border: none;
+      outline: none;
+      background-color: transparent;
+      font-size: 14px;
+      color: rgb(0, 143, 180);
+      margin: 0 10px;
+
+      &:disabled{
+        pointer-events: none;
+        color: rgb(187, 194, 220);
+      }
+    }
+  }
+}
