+++ src/pages/Browse/TaskDetail/Question/QuestionInput/index.js
@@ -1,0 +1,1 @@
+export { default } from './QuestionInput';
